RoomMessageEvent::fileNameToDownload()
Moved from Room::Private. Sonar for some reason insisted to pay
attention to this particular FIXME, well why not :)
